Extract base class from DeltaLakeHistoryTable
